--- a/benchmarks/benchmarks.py
+++ b/benchmarks/benchmarks.py
@@ -9,23 +9,22 @@ class TestMessage(betterproto.Message):
     baz: float = betterproto.float_field(2)
 
 class BenchMessage:
+    """Test creation and usage a proto message.
     """
-    An example benchmark that times the performance of various kinds
-    of iterating over dictionaries in Python.
-    """
+
     def setup(self):
         self.cls = TestMessage
         self.instance = TestMessage()
+        self.instance_filled = TestMessage(0, "test", 0.0)
 
     def time_overhead(self):
         """Overhead in class definition.
         """
-        for _ in range(100):
-            @dataclass
-            class Message(betterproto.Message):
-                foo: int = betterproto.uint32_field(0)
-                bar: str = betterproto.string_field(1)
-                baz: float = betterproto.float_field(2)
+        @dataclass
+        class Message(betterproto.Message):
+            foo: int = betterproto.uint32_field(0)
+            bar: str = betterproto.string_field(1)
+            baz: float = betterproto.float_field(2)
 
     def time_instantiation(self):
         """Time instantiation
@@ -50,6 +49,10 @@ class BenchMessage:
         self.instance.foo = 0
         self.instance.bar = "test"
         self.instance.baz = 0.0
+    
+    def time_serialize(self):
+        """Time serializing a message to wire."""
+        bytes(self.instance_filled)
 
 
 class MemSuite: