@cetanu 2020-09-11
Looks pretty cool, probably just need to add `asv` to poetry dependencies somewhere, 

and does this need to be included in the build process? Should the output be parsed and validated? I think I will wait for opinions on that