@adriangb 2020-09-11
Oh yeah forgot to add the dependency. It should be optional.

The way I've used benchmarks previously, they aren't run as part of CI, although they could be. Largely depends on how long they take to run. A decent option is for whoever is approving the PR to ask for benchmarks _if_ they suspect performance may be affected.