@nat-n 2020-09-13
Nice, this should help guide optimisation work. 

asv should be a dev dependency, and I don't like the look of changing the python version dependency, is this intentional?