@adriangb 2020-09-13
> asv should be a dev dependency, and I don't like the look of changing the python version dependency, is this intentional?

Added asv. That was not an intentional change, I reverted it.