@adriangb 2020-09-19
Is it normal for `project.lock` to change for a PR? Or should I not have committed that? I'm a relatively new Poetry user...