@nat-n 2020-09-19
> Is it normal for `project.lock` to change for a PR? Or should I not have committed that? I'm a relatively new Poetry user...

If you've added, removed or updated dependencies then yes the lockfile should also be updated to reflect this.